pr: handle feedback
