+++ wrapper-generator/src/main/kotlin/it/krzeminski/githubactions/wrappergenerator/GenerationEntryPoint.kt
@@ -1,10 +1,8 @@
 package it.krzeminski.githubactions.wrappergenerator
 
-import it.krzeminski.githubactions.wrappergenerator.domain.ActionCoords
-import it.krzeminski.githubactions.wrappergenerator.domain.WrapperRequest
 import it.krzeminski.githubactions.wrappergenerator.generation.generateWrapper
+import it.krzeminski.githubactions.wrappergenerator.generation.suggestDeprecations
 import it.krzeminski.githubactions.wrappergenerator.metadata.actionYmlUrl
-import it.krzeminski.githubactions.wrappergenerator.metadata.prettyPrint
 import java.nio.file.Paths
 
 /***
@@ -34,22 +32,6 @@ fun main() {
         |./gradlew ktlintFormat
     """.trimMargin()
     )
-}
-
-fun List<WrapperRequest>.suggestDeprecations(): String {
-    val nonDeprecatedCoords = this.map { it.actionCoords }
-        .filter { it.deprecatedByVersion == null }
-    val groupedBy: List<List<ActionCoords>> = nonDeprecatedCoords
-        .groupBy { coords -> "${coords.owner}/${coords.name}" }
-        .values.toList()
-    val messages = groupedBy.flatMap { list ->
-        val maxVersion = list.maxByOrNull { it.version }?.version ?: error("Unexpected empty list in $groupedBy")
-        list.filter { it.version != maxVersion }
-            .map { coords ->
-                """WARNING: newer version available for ${coords.prettyPrint}. Maybe add: deprecatedByVersion = "$maxVersion" ?"""
-            }
-    }
-    return messages.joinToString("\n")
 }
 
 private fun checkDuplicateWrappers() {
