+++ wrapper-generator/src/test/kotlin/it/krzeminski/githubactions/wrappergenerator/generation/DeprecationSuggestionsTest.kt
@@ -4,9 +4,8 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
 import it.krzeminski.githubactions.wrappergenerator.domain.ActionCoords
 import it.krzeminski.githubactions.wrappergenerator.domain.WrapperRequest
-import it.krzeminski.githubactions.wrappergenerator.suggestDeprecations
 
-class VersionSuggestionsTest : FunSpec({
+class DeprecationSuggestionsTest : FunSpec({
 
     val wrapperRequests: List<WrapperRequest> = listOf(
         WrapperRequest(
@@ -16,13 +15,13 @@ class VersionSuggestionsTest : FunSpec({
             ActionCoords("owner", "name", "v5"),
         ),
         WrapperRequest(
-            ActionCoords("owner", "other-name", "v1"),
+            ActionCoords("owner", "other-name", "v1", deprecatedByVersion = "v2"),
         ),
         WrapperRequest(
             ActionCoords("owner", "other-name", "v2", deprecatedByVersion = "v3"),
         ),
         WrapperRequest(
-            ActionCoords("owner", "other-name", "v3"),
+            ActionCoords("owner", "other-name", "v3", deprecatedByVersion = "v2"),
         ),
     )
 
@@ -30,6 +29,7 @@ class VersionSuggestionsTest : FunSpec({
         wrapperRequests.suggestDeprecations() shouldBe """
             WARNING: newer version available for ActionCoords("owner", "name", "v4"). Maybe add: deprecatedByVersion = "v5" ?
             WARNING: newer version available for ActionCoords("owner", "other-name", "v1"). Maybe add: deprecatedByVersion = "v3" ?
+            WARNING: ActionCoords("owner", "other-name", "v3") is deprecatedByVersion "v2", an older version ?
         """.trimIndent()
     }
 })
